changed type of latency to float64

+++ src/ILogParser.go
@@ -19,7 +19,7 @@ type HttpRequest struct {
   mobile         bool
   country        string
   city           string
-  latency        int
+  latency        float64
 }
 
 type ILogParser interface {
